+++ src/lib/components/module/ModuleCard.tsx
@@ -5,9 +5,8 @@ import { useMemo } from "react";
 import { AppLink } from "../AppLink";
 import { CustomIcon } from "../icon";
 import { AmpEvent, track } from "lib/amplitude";
-import type { IndexedModule } from "lib/services/move/moduleService";
 import { useVerifyModule } from "lib/services/move/moduleService";
-import type { BechAddr, Option } from "lib/types";
+import type { BechAddr, IndexedModule, Option } from "lib/types";
 
 import { CountBadge } from "./CountBadge";