+++ src/lib/pages/interact/index.tsx
@@ -10,12 +10,11 @@ import { ModuleSourceCode } from "lib/components/module";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
 import { useOpenNewTab } from "lib/hooks";
-import type { IndexedModule } from "lib/services/move/moduleService";
 import {
-  useAccountModules,
+  useModuleByAddressLcd,
   useVerifyModule,
 } from "lib/services/move/moduleService";
-import type { Addr, ExposedFunction } from "lib/types";
+import type { Addr, ExposedFunction, IndexedModule } from "lib/types";
 import { getFirstQueryParam } from "lib/utils";
 
 import {
@@ -90,27 +89,25 @@ export const Interact = () => {
   const moduleNameParam = getFirstQueryParam(router.query.moduleName);
   const functionNameParam = getFirstQueryParam(router.query.functionName);
   const functionTypeParam = getFirstQueryParam(router.query.functionType);
-  const { refetch } = useAccountModules({
+  const { refetch } = useModuleByAddressLcd({
     address: addressParam as Addr,
     moduleName: moduleNameParam,
-    functionName: functionNameParam,
     options: {
       refetchOnWindowFocus: false,
       enabled: false,
       retry: false,
       onSuccess: (data) => {
-        if (!Array.isArray(data)) {
-          setModule(data);
-          if (functionNameParam) {
-            const fn = [...data.viewFunctions, ...data.executeFunctions].find(
-              (exposedFn) => exposedFn.name === functionNameParam
-            );
-            if (fn) {
-              handleSetSelectedType(fn.is_view ? "view" : "execute");
-              setSelectedFn(fn);
-            }
-          } else if (functionTypeParam)
-            handleSetSelectedType(functionTypeParam);
+        setModule(data);
+        if (functionNameParam) {
+          const fn = data.parsedAbi.exposed_functions.find(
+            (exposedFn) => exposedFn.name === functionNameParam
+          );
+          if (fn) {
+            handleSetSelectedType(fn.is_view ? "view" : "execute");
+            setSelectedFn(fn);
+          }
+        } else if (functionTypeParam) {
+          handleSetSelectedType(functionTypeParam);
         }
       },
     },