+++ src/lib/types/move/module.ts
@@ -1,15 +1,30 @@
-import type { BechAddr } from "lib/types";
+import type { HexAddr } from "../addrs";
+import type { Nullable } from "../common";
+import type { Proposal } from "../proposal";
 
-export interface ModuleInfo {
+import type { ExposedFunction, InternalModule, ResponseABI } from "./abi";
+
+export interface IndexedModule extends InternalModule {
   // NOTE: can also be an ica or a contract
-  address: BechAddr;
-  name: string;
-  functions?: {
-    view: number;
-    execute: number;
-  };
+  address: HexAddr;
+  parsedAbi: ResponseABI;
+  viewFunctions: ExposedFunction[];
+  executeFunctions: ExposedFunction[];
+}
+
+export interface ModuleInfo
+  extends Pick<IndexedModule, "address" | "moduleName">,
+    Partial<Pick<IndexedModule, "viewFunctions" | "executeFunctions">> {
   height?: number;
   latestUpdated?: Date;
   isRepublished?: boolean;
   isVerified?: boolean;
+}
+
+export interface ModuleData extends IndexedModule {
+  recentPublishTransaction: Nullable<string>;
+  recentPublishProposal: Nullable<Pick<Proposal, "id" | "title">>;
+  recentPublishBlockHeight: number;
+  recentPublishBlockTimestamp: Date;
+  isRepublished: boolean;
 }